langchain_learning/Chatbot_using_groq/requirements.txt
langchain-core
langgraph>=0.2.28
langchain-groq
streamlit
python-dotenv
gradio